审核等关联记录修改

+++ dao/AccountDao.go
@@ -46,7 +46,7 @@ type UpdateAccountAmountArgs struct {
 	TaskId             int64
 	ShopId             int
 	OrderId            int
-	InOutId            int
+	InOutId            int64
 	Remark             string
 }
